feat: change layout structure and add semantic html to header and footer
