--- src/components/layout/index.module.css
@@ -1,3 +1,0 @@
-.container {
-    background-color: white;
-}
